Hierarchic selector loading indicator hide when loaded #7100
Previously the hierarchic selector did not update the loading indicators,
nor showed children indicators even though the network responded properly.
This lead end-user to believe that the server was slow and/or the interface
was buggy and unreliable.

This was because NaturalDropdownContainerComponent had
ChangeDetectionStrategy.OnPush which cascaded onto the hierarchic selector,
effectively disabling Angular automatic change detection behavior. That
strategy was enabled a long time ago in natural-search project without an
explicit reason. And it seems it is no longer necessary (if it ever was).

Delays were introduced in AnyService to expose those kind of bugs and better
reflect what a real world case would be with some network delays.