Support natural-select-*[required] #7292
With e757806 we lost support for [required] attribute on natural-selects. The input was only used to detect a warning.

But we need that [required] input.

This commit add support for it in a controlled way to match the ReactiveAsteriskDirective :

If template explicits the required attribute, it's used to determine the inner constraints. If not, the reactive forms validators are used.