Avoid intermediate Observable for getOne() #7739
Similar to 2c2c8752d145695f2b3c9934aa140e9ed6454c53 but for getOne().
And for the same reason too. We must be able to catch error in our apps
if needed, so we must avoid intermediate observable that do not properly
forward errors.

With this fix we finally restore the historical behavior of redirecting to
an error page when our app resolve an non-existing object.

Most likely, we should to the same for getAll/watchAll some day.