BREAKING Use service as generic #29
This simplifies usage in inheriting components, reducing copy-paste
errors, like so:

```diff
- export class ActionComponent extends AbstractDetail<Action['action'],
-     ActionVariables,
-     CreateAction['createAction'],
-     CreateActionVariables,
-     UpdateAction['updateAction'],
-     UpdateActionVariables,
-     DeleteActions['deleteActions']> {
+ export class ActionComponent extends AbstractDetail<ActionService> {
```

This also makes `this.service` strongly typed in inheriting components. Making
the types flow through the whole application.

Finally it also allows for Natural components to type their returned values. Typically
used un NaturalSelectComponent. So usage become strongly typed like so:

```html
<natural-select
    [service]="userService"
    (selectionChange)="selectionChange($event)"
></natural-select>
```

```diff
-public selectionChange(user: string | Literal | null): void {
+public selectionChange(user: string | Users_users_items | null): void {
     // do something
 }
```

One downside is that within Natural we become slightly more permissive. Typically
NaturalSelectComponent should be able to specify `unknown` for things that are not
actually used (eg: Tcreate) by this specific component. But we could not make it work
that way. And because this looseness does **not** leak to the app, it is tolerable for now.