service.create() may return `null` #7921
It may happen that the API returns `null` on a create mutation. In that
the service will return `null` instead of the input merged with result.

This is useful in Epicerio where orderService might not always create
an order, and we want the GUI to react accordingly. So we need a way
to make the distinction between `null` and non-`null` results.