BREAKING ModelService.watchAll does not fetch until subscribe #8174
It used to subscribe to Apollo (so fetch from network) whenever variables
changed, even if nobody subscribed to the result. That meant that we sent
XHR whose results was never usable by anybody.

Now we subscribe to Apollo whenever variables changes **and** if we have
an active subscription. This has the added benefit of properly forwarding
errors that Apollo may throw on its observables, and have a standard workflow
of unsubscribing (meaning potential for cancelling XHR).

This is breaking because the signature change, and user **must** unsubscribe
from outside.