Migrate from Protractor and Pupeteer to Playwright #8336
Protractor is EOL, and Playwright can replace entirely Pupeteer (for
our usage for Karma). This simplify a bit the maintenance since we
can install system deps automatically, and we are sure that everything
is always synced.

For now we only test on Chromium, because that's what we always did, but
we could also test other browsers later.