New `[naturalCustomCss]` directive #8782
This allows to apply dynamic CSS inside a component, with auto-scoping
withing that component.

`columnResizing()` is removed because it prevented to apply our custom
class within the editor. And column sizes would probably be better to
set via the new custom CSS. But if we really need it in the future, then
we should be able to use inherit from `TableView` and give our
implementation to `columnResizing()`.