natural-table-button fallbacks to span when no link is given #8905
In some edge case we don't have any links to link to. Eg: because the
viewer is not allowed to access the listed item. In this case we allow
to gracefully fallback on a simple span to still show the icon and/or
label, but not as button that would not do anything.