Don't show delete button on create page #8589
We used to show the delete button for a very short time between end of
creation and redirection to next page.

Now the button has a creation mode which is changed only when the page
changes. So even if the model receive an ID after creation, the button
will still stay in creation mode. The only way to change mode is to
navigate to another page.