BREAKING: Prevent mutation after logout #8398
It was possible that a debounced mutation happened after a logout. That
would lead to correct server error. But most importantly that means data
loss for the user who thought he saved something but didn't.

Now the debounced updates can be flushed immediately. So applications
should flush all pending updates just before logging out.

This unfortunately is a breaking change because a new injection is
required in all inheriting classes of NaturalAbstractModelService

A bit unrelated, but also rename AnyService to ItemService