NaturalPersistenceService can ignore invalid stored values #9187
If an app intentionally breaks the API, the app should have a way to
tell Natural which stored values are now invalid. This is done through
an optional global callback which is injected into the service.

If the validator returns false, then the service will ignore the value
and return `null` instead. The stored value will only be updated when
end-user make a new sort or filter. Then a new valid value will be
stored as usual. That means it is possible to see an URL which have no
impact on the page content. That is on purpose to allow power-users to
edit the URL manually without losing their changes between edits.