BREAKING `<natural-file>` requires explicit upload callback #8464
We used to accept a model service to transparently upload the file to
the server via `service.create()`, and then the parent component had the
freedom to do extra steps upon `(modelChange)`.

This was problematic if the human navigated away from the page while the
file was still uploading. In that case, the parent component would get
destroyed before the upload finished, and `(modelChange)` would never
get fired. That means it was impossible to upload-and-link-to-object in
an atomic manner. And it was quite likely that a human uploaded an image
that never got linked to anything and thus became totally unusable.

Our solution is to keep both mutations (creation and link) as close
together as possible with the new `[uploader]`. This way we can
(almost¹) guarantee that both mutations are called, even if the human is
navigating away quickly.

¹ unless if there is a server/network failure, then it still is possible
that an image is created but linked to nothing, but we hope it's much
less likely