BREAKING `<natural-icon>` is now `mat-icon[naturalIcon]`
We lose the possibility to set a label, but that was never used at all.
And instead we gain a much easier interoperability with `<mat-icon>`
special layouting done by Angular Material.

In most cases `[size]` probably should not be used anymore, since
Angular Material will do the work for us. But we keep it for corner
cases.