BREAKING natural-search is mobile friendly #8670
natural-search will adapt its layout to mobile screen by hiding some
buttons and repositioning others, in order to avoid horizontal
scrolling as much as possible.

This is a breaking change because natural-search no longer support
transcluded content. In most cases, you should not try to add things on
the right of natural-search, because it is not mobile friendly. Instead,
you should probably the new column picker, placed outside
natural-search.