Don't collapse to mobile mode if only column picker button #8670
Because this would not save any space, and it would make selecting
columns especially cumbersome via two sub-menus.

However, if there's no column choices at all, and there is a single
other button, then we still collapse, because we want to show the
textual label of the single button. We assume the column picker is
familiar enough to not show text, but all others should explicitly tell
what they are doing via textual description.