Never do anything with `null` relations
When a user adds a new selection, the `<natural-select>` will first
emit that selection, then `<natural-relations>` will add the relation,
then it will clear the selection of the `<natural-select>`, to be ready
for the next addition.

At this point `<natural-select>` will correctly emit `null`. So
`<natural-relations>` will receive `null` as a pseudo-relation to be
added. But this has to be ignored entirely. Because, first of all, it
obviously is not a valid relation for the server, and also because if
`<natural-relations>` try to again clear the selection of the
`<natural-select>`, the whole thing will enter an infinite loop.

Fixes #240