BREAKING `data.model` is synced with Apollo cache #9844
`NaturalAbstractModelService.resolve()` now resolves in two steps. First
it asks Apollo to fetch the object from network to force refresh Apollo
cache. Then only, the resolving is considered finished, and we emit an
observable that watches Apollo cache for the model. So the component can
receive the observable from the route, subscribe to it and immediately
receive an up-to-date model. And all subsequent updates to the model
will be emitted from Apollo cache too.

But the real breaking part is that previously the route resolved data
shape:

```ts
{
    permission: myPermissions,
    action: {
        model: myAction,
        statuses: myStatuses,
        priorities: myPriorities,
    }
}
```

Now, because `NaturalAbstractModelService.resolve()` only resolve the
model, it becomes the simpler:

```ts
{
    permission: myPermissions,
    model: myActionObservable,
    statuses: myStatuses,
    priorities: myPriorities,
}
```

So all overrides of `resolve()` must be migrate outside the model
service, and use the key `model` for the main model. Typically,
something like:

```ts
export const actionResolvers: ResolveData = {
    model: resolveAction,
    statuses: () => inject(NaturalEnumService).get('ActionStatus'),
    priorities: () => inject(NaturalEnumService).get('Priority'),
}
```