BREAKING drop `updatePartially()` in favor of `updateNow()` #9844
`updatePartially()` was created because `updateNow()` always injected
default values, so we could not update a single field at once.
But `updateNow()` should never have behaved that way, because all our
GraphQL mutations are entirely partial, and thus it is never required to
submit all fields for an update. I think `updateNow()` became this way
by accident when we introduced default value for creation (which is a
valid use-case), and/or because we had a few changes, very early on, on
how the API dealt with optional arguments.

So `updateNow()` does not inject default values anymore, and thus
`updatePartially()` loses its reason to exist.