Fallback to guesswork if browser does not give mime type #11606
For instance Chrome does not give a mime type for (some?) HEIC file.
Since we usually define `accept` via mimetype, we need introduce an
transparent, naive, transformation of accepted mimemtype into accepted
extensions. This is a best effort and does not work for all cases, but
it should work for most cases, and security is still guarantteed on
server side.