Support browsers that don't support `light-dark()` #11936
Some old browsers don't support `light-dark()`. That's Safari < v17.5.
That's roughly ~9.3% of our users.

We can tolerate that they won't be able to switch between light and dark
anymore, but we must at least provide light colors as a fallback to
show _something_.

So we use a technique similar to this:

```css
:root {
    --my-color: lightgray;
}

@supports (color: light-dark(white, black)) {
    :root {
        --my-color: light-dark(lightgray, darkgray);
    }
}

body {
    background: var(--my-color)
}
```